Refactor transformers to a package to expose just the items we want, and make maintenance easier.
